+++ braid/src/interface/mod.rs
@@ -30,8 +30,26 @@ use braid_data::{Datum, SummaryStatistics};
 
 /// Returns references to crosscat states
 pub trait HasStates {
+    /// Get a reference to the States
     fn states(&self) -> &Vec<State>;
+
+    /// Get a mutable reference to the States
     fn states_mut(&mut self) -> &mut Vec<State>;
+
+    /// Get the number of states
+    fn n_states(&self) -> usize {
+        self.states().len()
+    }
+
+    /// Get the number of rows in the states
+    fn n_rows(&self) -> usize {
+        self.states()[0].n_rows()
+    }
+
+    /// Get the number of columns in the states
+    fn n_cols(&self) -> usize {
+        self.states()[0].n_cols()
+    }
 }
 
 /// Returns and summrize data